Moved table wizard to action.wizard.table package
